Unify blog post card into one shared component
Both BlogPostsGrid and LatestNewsSection had drifted copy-paste `BlogPostCard`
implementations, and a third stale standalone `BlogPostCard.tsx` was dead. Replace
all three with a single shared component that takes the consumer's scoped CSS module
(preserving the distinct blog/landing visuals) plus a `headingLevel` prop, and uses
the correct i18n (`byAuthor`) and `useLocaleRoutes()` for both surfaces.
